Refactor: Change image ID type from Long to Int
This commit updates the `GetContentActivityResult` utility to use `Int` instead of `Long` for image identifiers.

This change affects the `GetContentActivityResult` class, its `launch` method, and the `rememberGetContentActivityResult` composable function, ensuring type consistency for the image ID across the component.

### Key Changes:

*   **`GetContentActivityResult.kt`**:
    *   The `imageId` property in `GetContentActivityResult` is now of type `MutableState<Int?>`.
    *   The `launch` method's `id` parameter has been changed from `Long?` to `Int?`.
    *   The `onSuccess` callback in `rememberGetContentActivityResult` now accepts an `Int?` instead of a `Long?`.

+++ common/core/activity/src/main/kotlin/kz/rymbek/platform/common/core/activity/GetContentActivityResult.kt
@@ -15,10 +15,10 @@ import org.koin.compose.koinInject
 class GetContentActivityResult(
     private val launcher: ManagedActivityResultLauncher<Uri, Boolean>,
     private val imagePath: MutableState<String?>,
-    private val imageId: MutableState<Long?>,
+    private val imageId: MutableState<Int?>,
     private val fileUtils: FileUtils,
 ) {
-    fun launch(id: Long? = null) {
+    fun launch(id: Int? = null) {
         val filePath = fileUtils.createFilePath()
         imagePath.value = filePath
         imageId.value = id
@@ -30,10 +30,10 @@ class GetContentActivityResult(
 fun rememberGetContentActivityResult(
     contract: ActivityResultContract<Uri, Boolean> = ActivityResultContracts.TakePicture(),
     fileUtils: FileUtils = koinInject(),
-    onSuccess: (String, Long?) -> Unit,
+    onSuccess: (String, Int?) -> Unit,
 ): GetContentActivityResult {
     val filePath = remember { mutableStateOf<String?>(null) }
-    val fileId = remember { mutableStateOf<Long?>(null) }
+    val fileId = remember { mutableStateOf<Int?>(null) }
 
     val launcher = rememberLauncherForActivityResult(
         contract = contract,
